requirements.txt
pandas>=2.0
numpy>=1.26
scipy>=1.11
plotly>=5.18
yfinance>=0.2.36
pyyaml>=6.0
pyarrow>=14.0
pytest>=7.4
